+++ l4/pkg/uclibc/lib/contrib/uclibc/libc/misc/internals/tempname.c
@@ -113,7 +113,7 @@ int attribute_hidden ___path_search (char *tmpl, size_t tmpl_len, const char *di
 	return -1;
     }
 
-    sprintf (tmpl, "%.*s/%.*sXXXXXX", dlen, dir, plen, pfx);
+    sprintf (tmpl, "%.*s/%.*sXXXXXX", (int)dlen, dir, (int)plen, pfx);
     return 0;
 }
